Move internal directory under root directory
Having 'internal' directory under 'pkg' directory is not accordingly to Golang
good coding habits, hence move 'internal' directory under root directory.
